ScrollAll() (#2443)
* ScrollAll()

Exposes a sliced scroll behind an observable with an optional max degree
of parallelism

* implemented review feedback from @russcam, unneccessary cast + only set scroll to _doc order if the user did not supply a sort of their own

+++ src/Nest/Document/Multiple/Reindex/ReindexResponse.cs
@@ -20,7 +20,7 @@ namespace Nest
 		/// <summary>
 		/// The no of scroll this result represents
 		/// </summary>
-		int Scroll { get; }
+		long Scroll { get; }
 
 		/// <summary>
 		/// Whether both the scroll and reindex result are valid
@@ -37,7 +37,7 @@ namespace Nest
 		public IBulkResponse BulkResponse { get; internal set; }
 		public ISearchResponse<T> SearchResponse { get; internal set; }
 
-		public int Scroll { get; internal set; }
+		public long Scroll { get; internal set; }
 
 		public bool IsValid =>
 			this.BulkResponse != null && this.BulkResponse.IsValid
